requirements.txt
streamlit
torch
torchvision
transformers
numpy
pillow
librosa
scikit-learn
